backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
asyncpg==0.29.0
sqlalchemy[asyncio]==2.0.30
bcrypt==4.2.0
python-jose[cryptography]==3.3.0
python-multipart==0.0.9
minio==7.2.7
pydantic-settings==2.2.1
pydantic[email]==2.7.1
python-pptx==0.6.23
lxml==5.2.1
